requirements.txt
Flask==3.1.2
requests==2.32.5
beautifulsoup4==4.13.5
python-dotenv==1.1.1
lxml==6.0.1
